Add Symfony6 compatibility
